Added support for nil and buffer types
Fleshed out the documentation by adding the exact value each approach / constant represents.

Added string approaches to assist in specialized cases where a string is <constant><number> or <number><constant>.  The generalized case is slowed down slightly, however the benefit outweighs the cost (due to how prevalent such an approach is used).  Often User_63461532053 is used rather than 63461532053 so that the developer can easily understand that 63461532053 is a UserId and not an ItemId or some other id.

Vector sizes have changed due to constant vectors and numbers existing in the design, so recalculations are necessary.