Adjusted table drastically to improve speed/memory
Tables have a higher effective buffer size, however will be much faster and cheaper to create.

The previous and current approach require counting the size for the array and for the dictionary, with the first few bytes being set once the table has finished being stored.  Prior we would need to shift the entire buffer a few bytes over for EVERY table, which is incredibly expensive and slow for the general case (alot of developers use tables to organize the format).  By choosing to avoid this overhead, we need to tank the potentially high cost of adding 0-5 bytes per table.  In most cases it will be 4-5, but for larger tables there will be less cost.

The cost should not be too harmful as the effective buffer size should still be less than the size of the output from JSONEncode in most cases.  The worst case used to be 33% worse for tables, but it is now 2x worse.

`{"A"}` becomes '["A"]' (5 bytes) for JSON but is now 10 bytes for BufferSerializer.  `{"A", "A"}` is `["A","A"]` in JSON (9 bytes) but only 13 bytes for BufferSerializer.  So the worst case only impacts tiny tables (we store "A" as 3 bytes but JSON stores it as 4).

As a side note, reducing the table op codes allows for the op codes to be reallocated to other sections.  Such as constants or for future aspects.