Reformatted vector
Vector and number 1 byte constants are now 32, the cost was that the 8 identity vectors now cost 2 bytes.  For scalar multiples of the identities, there are no changes.  The constants may be re-added, taking away 8 values from userdata approaches (since I pushed the amount to 10 from 6 using 4 of the regained values).

Future also gained 8 values, so if needed they can be repurposed.

We were able to reduce table from 28 to just 8 in the prior commit, so 8 went to number constants, 8 to future, and 4 to userdata.  8 came from vector to be put into vector constants.
I may choose to have more built-in constants, and this approach, forcing the value to be 2 bytes, is better for lower priority built-ins.  Like 2 or -1 or pi or huge or e or others.