Reverted vector and added userdata
8 approaches where taken from userdata and added to vector.

Userdata now supports only one custom approach.  Userdata is filtered first through the constants, and if the userdata is not a constant it gets passed to the function.

For tables:
If the value is effectively nil, meaning the function doesn't change the position associated with the buffer, then it will be overwritten or removed from the buffer and the key associated will be removed.  If there is no custom function specified, it will be treated as nil (assuming it isn't a constant).