Added performance expectations
I may use my personal format as a test, since it is reasonably complex enough for a sample.  For fair-use, constants WILL be used in serialize, BUT only for aspects that are constant or are known to prove beneficial (not making every value constant).

Note: The implementation isn't done, but there are a few tests that exist as proof of concept to prove that certain approaches are more optimal than others.  Also that highly specialized cases could warrant a wrapper, potentially achieving 100x in speed (if we can `buffer.fill` entire sections).