Added Perf results for serialize
As deserialize has yet to be implemented, I am not publishing the tests or code yet.

The serialize section can still be improved upon, and more in-depth tests need to be created.  I know that short strings are currently slower to process than longer strings, due to the concatenation approach.  Further tests need to wait until deserialize has been created though.

A `cheating` section was added to the tests as `serialize` uses NO constants and thus does not give a perfect reflection of practical usage cases for output size.  Both extremes, using all constants and using none, assist in realizing the midpoint between both approaches, which is the typical user's situation for the given structure.