Modified table
Table no longer stores the array part and dictionary part lengths and instead uses a terminating approach.  Much like null-terminating strings.  This allows for the minimal amount of bytes to be added whilst ensuring that the implementation does not cause performance changes.  I thought of multiple other approaches to keep the size, but they all produced some side-effect, such as the need to shift the entire buffer a few bytes or needing to keep track of the position of the byte(s) the hold the length or masking the length into the data.

For the worst case in a prior approach that kept the size, it would always be 7 bytes (for extremely large mixed tables).  However now the worst case is 3 bytes.

The performance section has been updated.  With serialize and cheating-serialize and their relevant compression aspects being slightly more expensive but producing smaller output sizes.  The example I am using only uses one table, so only 5 characters are saved, but in most cases many tables are used.  The new approach for tables allows for smaller tables to be more worthwhile and tables that contain many tables to be much cheaper.

For my personal approach without using constants (cheating) it made a minimal case go from 86 to 46 bytes, which was more than the modified JSON approach I was using prior which produced 56 bytes.  By cheating I was able to get it to 22 bytes.  Realistically the value would be 25 bytes though.

Here is the modified JSON representation: {"Equipments":[[1,["",[4,[0,0,0]]],[500],["Decal":[]]]]}.  The [0,0,0] portion represented the vector (0,0,0).

Note: For arrays it is recommended to have a nil value (like a function or thread) in between possible gaps between indexes in order to reduce the bytes used.  Preferably use the same function/thread so the table becomes faster to process.

Another Note: The memory / speed aspects are subject to change depending on the platform used.  I have @native enabled, which is basically a JIT, but not all platforms are supported by it, so performance could be slower.