Updated performance metrics
The dictionary portion of table deserialization is incomplete and the sample case uses an array, which is why it has yet to be published.

The way I am measuring speed and memory will need to change, as I only get the 50th percentile for speed and avg for memory.  The values keep fluctuating too much each time I test.

Serialization and deserialization speed can improve.  The current implementation relies on tables to store and pass the position, however I presume passing the values will be faster.  The plan is to optimize further AFTER it has been debugged and performance tested and published.  The userdata custom function aspect will not be included until the optimizations are complete (assuming the approach improves performance).

Side note: It is possible to select the position of the table you want to start the loop at, however it makes looping much slower.  I would have used this approach to separate arrays and dictionaries if the overhead wasn't so much (for extreme cases on both ends it could be 4+x faster or 4+x slower).
```luau
for i,v in next, tab, pos do
    -- ...
end
```