Improved string serialization perf
Strings that contain numbers like: Hello0 are no longer legible for the <const><num> approach.  The length of the <num> string must be more than 1.  The same goes for <num><const>.  This is due to "A0" going from 4 bytes (STRBYTE 2 A 0) to 5 bytes (CONCSN STRBYTE 1 A NUM0), meaning the other portion of the string must be a constant for any true benefit.

However <num> remains the same, as "0" will turn from 3 bytes to 2 bytes.

Also made Serialize ensure that all of the modules necessary were required prior to use.