Improved output size/perf for string conc
For strings with the first or last byte used to represent a number, only constant strings will be processed now.  This ensures that output size is improved and is 2x faster than the prior case for those values since they no longer need to be string serialized again.

For the general case, it maybe makes smaller strings 0.5% slower, however it seems to be relatively minimal.