Fixed nan vectors
Should a vector contain nan as any of its elements, it is not equal to itself!!  This is vital, else it would be impossible to store vector.create(0 / 0, 2, 4).

I ensured that vectors with nan now are cheaper, and that tables with nan vectors as values aren't added to the existing index table.