Added benchmark code and results
The benchmarking takes a minute due to the amount of tests and the results fluctuate depending on the platform.

I actually suggest overriding bench's print method so the results aren't as lengthy.  Maybe just printing the name, median for speed / avg for memory, and output.

Added a hint saying that I would provide more in-depth examples