Optimized number, buffer, and string
Honestly, very glad I set it up the way I did.  It is making this optimization so much easier for the simpler types.

Number is now 20% faster and 90% cheaper.
String is now 20% faster and 90% cheaper.
Buffer is now 20% faster and 90% cheaper.

Removed BufferData and PositionData types from Enumerator since they no longer serve a purpose.

The 90% cheaper is mainly because string / buffer were incredibly small.  If I used larger values to test, it would be not so drastic.