Patch serial
Fixed a few minor issues with string and table

There are some cases that take half as long and others that take around the same time as the prior version.  There is also another optimization that increases the gap slightly more by passing the buffer size.

Although a 10-50% speedup is nice, I have to still think about userdata.  I was expecting it to be at least 50% due to how heavily the old version was indexing tables.  The more inputs, the more confusing.  I can abstract away the complexity for userdata, but it adds overhead.