Pass size as well to optim
Serialization could be more performant for tables, however memory and speed have been improved a size-able amount.  Table serialization is also neater than the prior version, which slows down the execution slightly.

Grow has been removed, which ensures that buffer.len is not called and thus the overhead is not introduced.