Fleshed out the cframe example
The provided link showcases a cheaper approach, however vector / vector is alot easier to read/write and could produce similarish results.  The results, as noted in a comment, are an APPROXIMATION, they are very very close but sadly there is some error.  If I can find an approach that yields no error, I will update this using the approach.  The goal is for no loss of data, but I am also not willing to make it extremely expensive.

The provided link also has some error, moreso than mine since the 32bit floats (or 64 bit floats?) are converted into i16s.