Adjusting vector inflation and improved vector
When the buffer handles buffers, the best case no longer will need to inflate twice, however most cases still will.

Vector BYTE is now 5% faster.
Vector CHAR is now 4% slower
Vector TRYTE is now 2% slower
Vector NUMBER is now 10% faster best case and 2% faster worst case.  Number is the slowest of the 4 since it needs to call numSerializer three times, so the tradeoff is worth it.  By excluding bytes, we also ensure that an approach that is both faster and yields a smaller output size is run (BYTE should be prioritized when possible).  Char / tryte may be beneficial to support, although it'll slow them down.

Scalar and constants are unchanged.