Slight improvement to number
Surprisingly, double is 1% faster and float is 30% slower (since trim is called).  However!  For tables it is 2% faster.  So.. not that much of an improvement, but for number/vector heavy tables it is a much wanted speedup. :>

I am uncertain if inflating by the worst-case size scenerio is helpful or not.  For tables, it will never inflate twice in a row for vectors / numbers.  So no.