Added roundRobin to existing
Identified issue where extremely large tables may run indefinitely until exhaustion since the table cycle may be so deep.

I have not thought of many solutions yet, one is to, after filling existing, store all tables encountered.  This solution would require a value, like TABLE or ARRAY, probably 201 or what was TINY.

However, I have other thoughts on what we could use 201 for.  One of which is to allow for deserialization to be reworked by others to perform in parallel by storing the positions of each table at the end of the buffer, however this doesn't necessarily use 201.  So I guess it doesn't count completely.  The other use would be as a pointer to say the the current element of the table is equal to the prior element.  There are definitely more uses, and if I can find a way to not use 201 and keep it for later, I will.  I can bleed into my future approaches with the pointer to previous instead if needed.