Removal of string approaches
Removed the concatsn string approach.
Removed the concatns string approach.
Removed the num string approach.

Reasoning: All approaches MUST provide tangible benefit for the general case.  The purpose of these approaches was to make niche cases, "User_129493" much cheaper to store.  The purpose has been recognized as not enough to warrant the performance cost on all other string values, as the format designer can redesign to avoid usages of these aspects or use a table to store the ids.

For specific reasons.  Num was removed as the format designer should have used a number instead of a string if they wanted it to be one.  Concat number string and concat string number were removed since the benefit was outweighed by the cost, as described below.

Looking at the diff of compare_results, we can observe that serialization of this aspect of string only reduced dataset output sizes by the order of hundreds even for relatively large datasets.  We can also observe the removal caused serialization to be 25% faster for string intensive tables.

The readability and thus maintainability of string improved drastically due to the removal.