Added an example for Roblox userdata
The example provided does NOT include all Roblox userdata types.  The example does NOT take advantage of common constants.  The example does NOT supply the best approach for each userdata.

Most userdata should be backed by multiple approaches.  Also the developer should understand which userdata is a lightuserdata, that is if you create an EXACT replica of the item it will equal the replica.  Like vector.create(1, 2, 3) == vector.create(1, 2, 3).  This information is VITAL when supporting certain constants.  I would have added support for table constants if {500} == {500}.  Always check!