Updated README
Moved the majority of tecnical details into FORMAT.md.

Added a link in roblox-userdata to be able to find rose.

Changed how requiring BufferSerializer is done.  It may be changed again.  init.luau and @self are currently too complex for me to understand, so I may later adjust `./BufferSerializer/BufferSerializer` to just be `./BufferSerializer`.