Fixed cyclic table deserialization
The performance improvement seems unrealistic, 50%+ improvement to table deserialization for cases where table use is heavy.

I had a suspicion that deserialization was returning nil, but that is not the case.  How peculiar.  Most of the performance improvement SHOULD come from @native being able to now say that data IS a table.

compare_results and component_results where updated again.  In component_results we see no difference in performance for small tables.

Fixed a type error regarding regarding serialSwitch.