Adjusted the benchmarking environment
The benchmarking environment prior was performed in Roblox Studio using a near template game.  Just a single part floating in space and my character.  The benchmarking started immediately, which conflicted with the character creation phase and setup phase.  As such, I added a wait at the top of the script.

Another potential aspect that could impact the performance would be the physics engine, as such I changed the character into a model with a single anchored part.

These results show that the startup of the testing environment was interfering with the results, albeit slightly.