Request for connecttable to be implemented
Reasoning: ConnectN allows for the array section of tables to produce smaller output sizes, assuming the prior value is a table.  That is: { {1}, {1} } will become 1 byte shorter.  Most datasets can benefit from the addition of this feature, slightly.

Pros: Slightly reduced output for most datasets
Cons: 3 future approach slots are consumed.

DROP IF:
As the feature will not reduce table output sizes by much, by 0.1%-1%, the slowdown cannot be more than 8% for deserialize (due to the @native optimization) OR 4% for serialize.