Cleaned up table to improve readability
There are still points to improve regarding readability and some aspects have regressed.

The following changes only apply to tables:

Deserialize now uses a switch statement as well, although not as readable as serialize's.

Deserialize and serialize now have an existingCache function that handles that processing of adding elements to the existing cache.

Deserialize and serialize no longer require removal of empty tables from the existing cache, there were slight performance implications for general tables.

Deserialize and serialize now have a fast path for empty tables, as serialTable / deserialTable have removed all mention of empty tables.

In serialize, dictionaries are no longer assigned at the beginning and overwritten when incorrect.  All table type writings are performed after all knowledge regarding the table is held.

In serialize for arrays, the type of each element is obtained prior to the serialCache call for a potential later implementation of a feature.

Due to the above statement, the readability of serialCache was improved and nilSerialize [functions / threads] are faster to process.

Performance implications: Serialize is 1% slower and deserialize is 4% slower.