Split future support to reserve extender support
Of the 32 reserved values for future approaches, 16 will be allocated to extenders.  Extenders take on the risk of needing to break backwards compatibility and future compatibility in order to meet certain requirements.

Updated the performance section to hopefully point the links to the correct location.

The purpose section of BufferSerializer is quite bloated, but it can be worked on later.