Add short string approaches from MessagePack
Repurposed bytes that will not be used.  Buffer unknown would not be used, the same goes for number and most of vector.  Vector will still have a few just in case.

Strings of length 15- will now be 1 byte shorter.

Completely breaks backwards compatibility, as such will need to be a major release.