parallel.luau can load files
The loaded files cannot then load more files yet.  I have done it before in Java for Luau, and it is quite a pain.  We need to store the path to the file (and as we go further the path gets incredibly length).  We could and should resolve, but Lune doesn't have that built-in.

Signed-off-by: Noah Winn <111249648+nwinn-student@users.noreply.github.com>