Added a cores option, removing no-parallel and process-per-core
The default, assuming the --cores option isn't used, is to run CORES-1 processes at once.  Where CORES is assumed to be defined in process.env.

Fixed a bug where adding multiple luau options would cause the generated scripts to fail.

Fixed a bug where the debug option wouldn't work when set, and would say it is an invalid command.

Signed-off-by: Noah Winn <111249648+nwinn-student@users.noreply.github.com>