Closes #36
Added the ability to skip pairs for serialization.  Could be useful for extenders, but is required for table.

For table serialization, added constants that are paired will also have their actual value stored after the first time the value is serialized.

For table deserialization, we now need to store the pairs, and avoid table indexing as much as possible.

Signed-off-by: Noah Winn <111249648+nwinn-student@users.noreply.github.com>